requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.0
requests==2.31.0
web3==7.10.0
openai==1.3.5
gunicorn==21.2.0